+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0.2] - 2025-10-27
+
+### Versions
+
+- CSC FastAPI Backend:  0.4.0.0
+- CSC React Frontend:   0.4.0.2
+- CSC Grasshopper Interface: 0.4.0.0
+
+### Changed
+
+#### CSC React Frontend
+
+- **QR Code Scanning**: Modularized QR Code Scanner
+
 ## [0.4.0.1] - 2025-10-27
 
 ### Versions